+++ database/seeds/RecruiterDatabaseSeeder.php
@@ -1,0 +1,12 @@
+<?php
+
+use Illuminate\Database\Seeder;
+use JobApis\JobsToMail\Models\Recruiter;
+
+class RecruiterDatabaseSeeder extends Seeder
+{
+    public function run()
+    {
+        // Seed recruiting companies here
+    }
+}