+++ resources/views/searches/index.blade.php
@@ -8,11 +8,14 @@
 
 <div class="row searches">
     <div class="col-lg-8 offset-lg-2 col-sm-12 offset-sm-0">
+        <div class="card card-block" style="margin: 30px 0;">
+            @include('advertisements.jobs-hub')
+        </div>
+
         <h3>Your Searches</h3>
         @foreach($searches as $search)
             <div class="card card-block">
-                <h4 class="card-title">"{{ $search->keyword }}"</h4>
-                <p>Location: "{{ $search->location }}"</p>
+                <h4 class="card-title">"{{ $search->keyword }} in {{ $search->location }}"</h4>
                 @include('searches.components.search-dates')
                 @include('searches.components.no-recruiters-checkbox')
                 @include('searches.components.action-links')