+++ tests/unit/Filters/JobFilterTest.php
@@ -58,7 +58,7 @@ class JobFilterTest extends TestCase
         $results = $this->filter->sort($jobsArray, 10, 50);
 
         // Ensure that the results cut off at the max
-        $this->assertLessThan(20, count($results));
+        $this->assertLessThanOrEqual(20, count($results));
 
         // Ensure that each job has an earlier date than the one previous
         $previousDate = null;